Cleanup TestKNNOperator, collapse unit test into knn_test.go (#1917)
